refactor: 커스텀 예외 적용

+++ src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,6 +1,10 @@
 package nextstep.schedule;
 
 import java.util.List;
+import nextstep.error.ErrorCode;
+import nextstep.error.exception.RoomReservationException;
+import nextstep.reservation.Reservation;
+import nextstep.reservation.ReservationDao;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -9,10 +13,12 @@ import org.springframework.stereotype.Service;
 public class ScheduleService {
     private ScheduleDao scheduleDao;
     private ThemeDao themeDao;
+    private ReservationDao reservationDao;
 
-    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
+    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao, ReservationDao reservationDao) {
         this.scheduleDao = scheduleDao;
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     public Long create(ScheduleRequest scheduleRequest) {
@@ -25,6 +31,10 @@ public class ScheduleService {
     }
 
     public void deleteById(Long id) {
+        List<Reservation> reservations = reservationDao.findByScheduleId(id);
+        if (reservations.size() > 0) {
+            throw new RoomReservationException(ErrorCode.SCHEDULE_CANT_BE_DELETED);
+        }
         scheduleDao.deleteById(id);
     }
 }
